pyproject.toml
[project]
name = "repl-nix-enhancedflashcardsystem"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.36",
    "openai>=1.57.2",
    "python-dotenv",
    "flask-migrate>=4.0.7",
    "aiohttp>=3.11.10",
    "hypercorn>=0.17.3",
    "asgiref>=3.8.1",
    "flask-cors",
    "alembic>=1.14.0",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.0",
    "numpy>=2.2.0",
    "tensorflow>=2.14.0",
    "transformers>=4.47.0",
    "torch>=2.5.1",
    "werkzeug",
    "flask-limiter>=3.9.2",
    "pyjwt>=2.10.1",
    "prometheus-client>=0.21.1",
    "flask-caching>=2.3.0",
    "redis>=5.2.1",
    "flask-sock>=0.7.0",
    "flask-prometheus-metrics>=1.0.0",
    "psutil>=6.1.0",
    "pyyaml>=6.0.2",
    "pydantic",
    "flask-wtf>=1.2.2",
    "flask-security>=5.5.2",
    "gunicorn>=23.0.0",
    "pytest",
    "backend>=0.2.4.1",
    "requests",
    "jsonschema>=4.23.0",
]
